requirements.txt
arxiv==1.4.2
PyYAML==6.0
slack-sdk==3.23.0
